Major rewrite of install-spark and supporting scripts in order to support a central configuration file. New arguments to install-spark include "-c <config_file_on_s3>" and "-l <spark_log_threshold>". Also includes Spark 1.2.1 release as build version 1.2.1.a.
Reorganization of repo to support continual developement of scripts.
