+++ spark/config.file
@@ -1,0 +1,46 @@
+#Config file is a tab-delimited file with these expected fields -
+# field1 - EMR AMI version (may be a, a.b or a.b.c format or 'default', default will be used if exact match cannot be found)
+# field2 - Spark build version (the REQUSTED_VERSION is matched against this field, if not specified it will look for the term 'default', default will be used if exact match cannot be found)
+# field3 - (optional) shell program to use to execute the install script, defaults to python
+# field4 - s3 location of a specific install script for this AMI and Spark version (is ran with environment variables S3SparkInstallPath, SparkBuild and Ec2Region)
+# field5 - s3 path to files that will be used for Spark binaries, how this field is used is up to the install script, sets enviroment variable 'SparkS3InstallPath'
+# field6 - (optional) s3 path to the script executed when "-x" argument is used, defaults to AWS provided version if empty
+# field7 - (optional) s3 path to the script executed when "-g" argument is used to install ganlia, defaults to AWS provided version if empty
+# field8 - (optional) s3 location for Scala binaries to install, how this field is used is up to the install script, set enviroment variable 'ScalaS3Location'
+#
+default	default	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.2.1/spark-1.2.1.a.tgz	s3://support.elasticmapreduce/spark/maximize-spark-default-config	s3://support.elasticmapreduce/spark/install-ganglia-metrics	s3://support.elasticmapreduce/spark/scala/scala-2.10.3.tgz
+#experimental builds
+default	1.2.1.a-2.11	python	s3://support.elasticmapreduce/spark/scala-2.11-builds/install-spark-v2-scala-2.11.py	s3://support.elasticmapreduce/spark/scala-2.11-builds/spark-1.2.1.a.tgz
+default	1.3	python	s3://support.elasticmapreduce/spark/install-spark-1.x.py	s3://support.elasticmapreduce/spark/1.3/
+default	1.2	python	s3://support.elasticmapreduce/spark/install-spark-1.x.py	s3://support.elasticmapreduce/spark/1.2/
+default	1.1	python	s3://support.elasticmapreduce/spark/install-spark-1.x.py	s3://support.elasticmapreduce/spark/1.1/
+# older configs
+3.3	default	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.2.1/spark-1.2.1.a.tgz
+3	1.2.1.a	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.2.1/spark-1.2.1.a.tgz
+3	1.2.0.a	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.2.0/spark-1.2.0.a.tgz
+3	1.1.1.e	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.1.1/spark-1.1.1.e.tgz
+3	1.1.1.d	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.1.1/spark-1.1.1.d.tgz
+3	1.1.1.c	python	s3://support.elasticmapreduce/spark/install-spark-script.py	s3://support.elasticmapreduce/spark/1.1.1/spark-1.1.1.c.tgz
+3	1.1.1.b	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.1/spark-1.1.1.b.tgz
+3	1.1.1.a	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.1/spark-1.1.1.a.tgz
+3.2	default	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.tgz
+3.2.1	default	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.before323.tgz
+3.2.0	default	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.before323.tgz
+3.2.1	1.1.0.h	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.before323.tgz
+3.2.0	1.1.0.h	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.before323.tgz
+3.1	default	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.tgz
+3	1.1.0.h	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.h.tgz
+3	1.1.0.g	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.g.tgz
+3	1.1.0.f	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.f.tgz
+3	1.1.0.e	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.e.tgz
+3	1.1.0.d	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.c.tgz
+3	1.1.0.c	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.c.tgz
+3	1.1.0.b	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.b.tgz
+3	1.1.0	python	s3://support.elasticmapreduce/spark/install-spark-script-v2-before-hadoop-provided.py	s3://support.elasticmapreduce/spark/1.1.0/spark-1.1.0.tgz
+3	1.0.2	python	s3://support.elasticmapreduce/spark/ami-3.1.1/install-spark-1.0.2.py
+3.0	default	ruby	s3://support.elasticmapreduce/spark/1.0.0/install-spark-shark-yarn.rb-v2
+3.0	1.0.0	ruby	s3://support.elasticmapreduce/spark/1.0.0/install-spark-shark-yarn.rb-v2
+2.4	default	bash	s3://elasticmapreduce/samples/spark/0.8.1/install-spark-shark.sh
+2.4	0.8.1	bash	s3://elasticmapreduce/samples/spark/0.8.1/install-spark-shark.sh
+2.3	default	bash	s3://elasticmapreduce/samples/spark/0.8.1/install-spark-shark.sh
+2.3	0.8.1	bash	s3://elasticmapreduce/samples/spark/0.8.1/install-spark-shark.sh
